Analysis changes

+++ dqm/cscmonitor/include/emu/dqm/cscmonitor/i2oEmuMonitorMsg.h
@@ -19,7 +19,7 @@ typedef struct _I2O_EMU_DATA_MESSAGE_FRAME
   uint32_t  runNumber;
   uint32_t  runStartUTC;
   uint32_t  nEventCreditsHeld;
-  uint32_t errorFlag; // for DQM; filled by EmuRUI's server only
+  uint16_t  errorFlag; // for DQM; filled by EmuRUI's server only
 } I2O_EMU_DATA_MESSAGE_FRAME, *PI2O_EMU_DATA_MESSAGE_FRAME;
 
 #endif
